backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
opencv-python==4.8.1.78
transformers==4.35.2
torch
torchvision
reportlab==4.0.7
web3==6.11.3
python-dotenv==1.0.0
sqlalchemy==2.0.23
aiofiles==23.2.1
